requirements.txt
numpy==1.19.1
Pillow==7.2.0
Pygments==2.6.1
Pypubsub==4.0.3
pywin32==228
robotframework==3.1.2
robotframework-pythonlibcore==2.1.0
robotframework-ride==1.7.4.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.5.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.10
wxPython==4.0.7
